# docs/api-contracts/admin-module-openapi.yaml
...
openapi: 3.0.3
info:
  title: Admin Module API
  description: |
    Admin Module — управляющий модуль системы Artstore.

    **Admin Module не является auth-сервером.** Аутентификация и выдача JWT
    полностью делегированы внешнему Identity Provider (Keycloak). Admin Module
    получает JWT от API Gateway, проверяет claims (роли, scopes) и принимает
    решения по авторизации для своих endpoints.

    ## Аутентификация

    Все запросы (кроме Health) проходят через API Gateway, который:
    - Проверяет подпись JWT через JWKS Keycloak
    - Проверяет expiration
    - Форвардит claims в заголовках к Admin Module

    ### Admin Users (H2M)
    Аутентификация через Keycloak OIDC (Authorization Code + PKCE).
    JWT содержит `sub`, `realm_access.roles`, `groups`.

    Роли определяются маппингом Keycloak groups → Artstore roles:
    - `artstore-admins` → `admin` (полный доступ)
    - `artstore-viewers` → `readonly` (только чтение)

    Admin Module может **дополнить** (но не понизить) роль через
    локальные role overrides.

    ### Service Accounts (M2M)
    OAuth 2.0 Client Credentials flow через Keycloak.
    JWT содержит `client_id`, `scope`.

    Scopes:
    - `files:read` — чтение метаданных и скачивание файлов
    - `files:write` — загрузка, обновление и удаление файлов
    - `storage:read` — чтение информации о Storage Elements
    - `storage:write` — управление SE (sync, mode transition)
    - `admin:read` — чтение административных данных
...

    ## Shared PostgreSQL
...

...

...

...
servers:
...
paths:
...
  /api/v1/storage-elements/discover:
    post:
      tags: [storage-elements]
      summary: Предпросмотр Storage Element
      description: |
        Запрашивает информацию о SE по указанному URL (вызывает
        `GET /api/v1/info` на SE) и возвращает результат.
        **Не создаёт запись** — только предпросмотр.

        Доступно ролям `admin` и `readonly`.
      operationId: discoverStorageElement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscoverRequest"
      responses:
        "200":
          description: Информация о SE получена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoverResponse"
        "400":
          description: Некорректный URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          description: SE недоступен или вернул ошибку
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: SE_UNAVAILABLE
                  message: "Не удалось подключиться к Storage Element"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/storage-elements:
    post:
      tags: [storage-elements]
      summary: Регистрация Storage Element
      description: |
        Регистрирует новый SE в системе. При регистрации автоматически:
        1. Запрашивает `GET /api/v1/info` на SE
        2. Запускает **full sync файловых метаданных**

        Доступно только роли `admin`.
      operationId: createStorageElement
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageElementCreate"
      responses:
        "201":
          description: SE зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageElement"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: SE с таким URL уже зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: CONFLICT
                  message: "Storage Element с таким URL уже зарегистрирован"
        "502":
          description: Не удалось получить информацию от SE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"

    get:
      tags: [storage-elements]
      summary: Список Storage Elements
      description: |
        Возвращает пагинированный список зарегистрированных SE.
        Используется Admin UI, Ingester и Query Module.

        Доступно: роли `admin`, `readonly`, SA с scope `storage:read`.
      operationId: listStorageElements
      security:
        - bearerAuth: [storage:read]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: mode
          in: query
          description: Фильтр по режиму работы
          schema:
            type: string
            enum: [edit, rw, ro, ar]
        - name: status
          in: query
          description: Фильтр по статусу
          schema:
            type: string
            enum: [online, offline, degraded, maintenance]
      responses:
        "200":
          description: Список SE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageElementListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/storage-elements/{id}:
    get:
      tags: [storage-elements]
      summary: Получить Storage Element
      description: |
        Возвращает данные SE по ID.
        Доступно: роли `admin`, `readonly`, SA с scope `storage:read`.
      operationId: getStorageElement
      security:
        - bearerAuth: [storage:read]
      parameters:
        - $ref: "#/components/parameters/StorageElementId"
      responses:
        "200":
          description: Данные SE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageElement"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [storage-elements]
      summary: Обновить Storage Element
      description: |
        Обновление данных SE (name, url).
        Mode, status, capacity обновляются через sync.
        Доступно только роли `admin`.
      operationId: updateStorageElement
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StorageElementId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StorageElementUpdate"
      responses:
        "200":
          description: SE обновлён
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageElement"
        "400":
          description: Некорректный запрос
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [storage-elements]
      summary: Удалить Storage Element
      description: |
        Удаление SE из реестра. Физические файлы на SE не удаляются.
        Доступно только роли `admin`.
      operationId: deleteStorageElement
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/StorageElementId"
      responses:
        "204":
          description: SE удалён из реестра
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/storage-elements/{id}/sync:
    post:
      tags: [storage-elements]
      summary: Синхронизация Storage Element
      description: |
        Полная синхронизация SE:
        1. `GET /api/v1/info` → обновить mode, status, capacity
        2. Постраничный `GET /api/v1/files` → синхронизировать file_registry

        `*.attr.json` на SE — единственный источник истины.

        Доступно роли `admin` и SA с scope `storage:write`.
      operationId: syncStorageElement
      security:
        - bearerAuth: [storage:write]
      parameters:
        - $ref: "#/components/parameters/StorageElementId"
      responses:
        "200":
          description: SE синхронизирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          description: Не удалось подключиться к SE
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error:
                  code: SE_UNAVAILABLE
                  message: "Не удалось подключиться к Storage Element"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Files Registry (5 endpoints)
  # =========================================================================
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT от Keycloak. Валидация подписи — на API Gateway.
        Admin Module проверяет claims (роли, scopes).

        Для Admin Users: `realm_access.roles`, `groups`
        Для Service Accounts: `scope`

  # =========================================================================
  # Параметры
  # =========================================================================
  parameters:
    Limit:
      name: limit
      in: query
      description: Количество записей на странице
      schema:
        type: integer
        minimum: 1
        maximum: 1000
        default: 100
    Offset:
      name: offset
      in: query
      description: Смещение от начала списка
      schema:
        type: integer
        minimum: 0
        default: 0
...
    StorageElementId:
      name: id
      in: path
      required: true
      description: UUID Storage Element
      schema:
        type: string
        format: uuid
...
  responses:
    Unauthorized:
      description: Отсутствует или невалидный JWT token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: "Требуется аутентификация"
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: "Недостаточно прав для выполнения операции"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: "Ресурс не найден"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: INTERNAL_ERROR
              message: "Внутренняя ошибка сервера"

  # =========================================================================
  # Схемы данных
  # =========================================================================
  schemas:
...
    StorageElement:
      type: object
      description: Зарегистрированный Storage Element
      required:
        - id
        - name
        - url
        - storage_id
        - mode
        - status
        - capacity_bytes
        - used_bytes
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          example: 770e8400-e29b-41d4-a716-446655440002
        name:
          type: string
          example: Moscow SE 01
        url:
          type: string
          format: uri
          example: https://se-moscow-01.kryukov.lan:8010
        storage_id:
          type: string
          description: Идентификатор SE (из SE /api/v1/info)
          example: se-moscow-01
        mode:
          type: string
          enum: [edit, rw, ro, ar]
          example: rw
        status:
          type: string
          enum: [online, offline, degraded, maintenance]
          example: online
        capacity_bytes:
          type: integer
          format: int64
          example: 107374182400
        used_bytes:
          type: integer
          format: int64
          example: 53687091200
        available_bytes:
          type: integer
          format: int64
          example: 53687091200
        last_sync_at:
          type: string
          format: date-time
          nullable: true
          description: Время последней синхронизации info
        last_file_sync_at:
          type: string
          format: date-time
          nullable: true
          description: Время последней синхронизации файлов
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    StorageElementCreate:
      type: object
      required:
        - name
        - url
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          example: Moscow SE 01
        url:
          type: string
          format: uri
          example: https://se-moscow-01.kryukov.lan:8010
    StorageElementUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        url:
          type: string
          format: uri
      minProperties: 1
    DiscoverRequest:
      type: object
      required:
        - url
      properties:
        url:
          type: string
          format: uri
          example: https://se-moscow-01.kryukov.lan:8010
    DiscoverResponse:
      type: object
      description: Информация от SE (GET /api/v1/info)
      required:
        - storage_id
        - mode
        - status
        - version
        - capacity
      properties:
        storage_id:
          type: string
          example: se-moscow-01
        mode:
          type: string
          enum: [edit, rw, ro, ar]
        status:
          type: string
          enum: [online, offline, degraded, maintenance]
        version:
          type: string
          example: 1.0.0
        capacity:
          type: object
          properties:
            total_bytes:
              type: integer
              format: int64
            used_bytes:
              type: integer
              format: int64
            available_bytes:
              type: integer
              format: int64
    SyncResponse:
      type: object
      description: Результат синхронизации SE
      required:
        - storage_element
        - file_sync
      properties:
        storage_element:
          $ref: "#/components/schemas/StorageElement"
        file_sync:
          type: object
          required:
            - files_on_se
            - files_added
            - files_updated
            - files_marked_deleted
            - started_at
            - completed_at
          properties:
            files_on_se:
              type: integer
              example: 1523
            files_added:
              type: integer
              example: 5
            files_updated:
              type: integer
              example: 12
            files_marked_deleted:
              type: integer
              example: 3
            started_at:
              type: string
              format: date-time
            completed_at:
              type: string
              format: date-time
    StorageElementListResponse:
      type: object
      required:
        - items
        - total
        - limit
        - offset
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/StorageElement"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        has_more:
          type: boolean

    # -----------------------------------------------------------------------
    # Files Registry
    # -----------------------------------------------------------------------
...
    ErrorResponse:
      type: object
      description: Стандартный формат ошибки (единый для всей системы Artstore)
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: |
                Машиночитаемый код ошибки:
                - `VALIDATION_ERROR` — некорректные входные данные
                - `NOT_FOUND` — ресурс не найден
                - `UNAUTHORIZED` — требуется аутентификация
                - `FORBIDDEN` — недостаточно прав
                - `CONFLICT` — конфликт (дублирующийся ресурс)
                - `SE_UNAVAILABLE` — SE недоступен
                - `IDP_UNAVAILABLE` — Keycloak недоступен
                - `INTERNAL_ERROR` — внутренняя ошибка
              example: NOT_FOUND
            message:
              type: string
              description: Человекочитаемое описание
              example: Ресурс не найден